requirements.txt
win32core==221.36
win32gui==221.6